+++ src/main/java/com/github/antag99/retinazer/EngineService.java
@@ -1,12 +1,25 @@
 package com.github.antag99.retinazer;
 
-/** General {@link Engine} service.
+/**
+ * General {@link Engine} service.
+ *
  * @see WireResolver
- * @see EntitySystem */
+ * @see EntitySystem
+ */
 public interface EngineService {
 
-	/** Called after the service is {@link Wire}d up.
-	 * If overriding, <b>ALWAYS</b> call {@code super} first. */
-	default void initialize() {}
+	/**
+	 * Called after the service is {@link Wire}d up.
+	 * If overriding, <b>ALWAYS</b> call {@code super} first.
+	 */
+	default void initialize() {
+	}
 
+	/**
+	 * Called on each {@link Engine#update}.
+	 *
+	 * @param delta time in seconds since last update
+	 */
+	default void update(float delta) {
+	}
 }