+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `Engine.getServices` to retrieve service subclasses
 - Made `FamilyPresenceWatcherSystem` more open for extensions
 - Split existing `EntityProcessorSystem` functionality into separate subclass `Single`, similar to how `FamilyPresenceWatcherSystem` handles things
+- Lift `update` into `EngineService` - all services can now receive updates
 
 # Version 0.2.5 (released 2020-01-21)
 - `EngineService`s are now automatically wired and initialized